move some messages into common module to get rid of node dependency in encoder
